Add Redis optimization for MinMax normalization

+++ python/preprocessing/MinMaxScaler.py
@@ -8,18 +8,25 @@ from serialization import LambdaThread, get_all_keys
 from threading import Thread
 
 class LocalBounds(LambdaThread):
-    def __init__(self, s3_bucket_input, s3_key):
+    def __init__(self, s3_bucket_input, s3_key, redis):
         Thread.__init__(self)
+        redis_s = "1"
+        if not redis:
+            redis_s = "0"
         self.d = {
             "s3_bucket_input": s3_bucket_input,
             "s3_key": s3_key,
             "action": "LOCAL_BOUNDS",
-            "normalization": "MIN_MAX"
+            "normalization": "MIN_MAX",
+            "redis": redis_s
         }
 
 class LocalScale(LambdaThread):
-    def __init__(self, s3_bucket_input, s3_key, s3_bucket_output, lower, upper):
+    def __init__(self, s3_bucket_input, s3_key, s3_bucket_output, lower, upper, redis):
         Thread.__init__(self)
+        redis_s = "1"
+        if not redis:
+            redis_s = "0"
         self.d = {
             "s3_bucket_input": s3_bucket_input,
             "s3_key": s3_key,
@@ -27,10 +34,11 @@ class LocalScale(LambdaThread):
             "action": "LOCAL_SCALE",
             "min_v": lower,
             "max_v": upper,
-            "normalization": "MIN_MAX"
+            "normalization": "MIN_MAX",
+            "redis": redis_s
         }
 
-def MinMaxScaler(s3_bucket_input, s3_bucket_output, lower, upper, objects=[], dry_run=False):
+def MinMaxScaler(s3_bucket_input, s3_bucket_output, redis, lower, upper, objects=[], dry_run=False):
     s3_resource = boto3.resource("s3")
     if len(objects) == 0:
         # Allow user to specify objects, or otherwise get all objects.
@@ -43,7 +51,7 @@ def MinMaxScaler(s3_bucket_input, s3_bucket_output, lower, upper, objects=[], dr
         while len(b_threads) > 400:
             t = b_threads.popleft()
             t.join()
-        l = LocalBounds(s3_bucket_input, i)
+        l = LocalBounds(s3_bucket_input, i, redis)
         l.start()
         b_threads.append(l)
 
@@ -52,40 +60,40 @@ def MinMaxScaler(s3_bucket_input, s3_bucket_output, lower, upper, objects=[], dr
 
     start_global = time.time()
     print("LocalBounds took {0} seconds...".format(start_global - start_bounds))
+    if redis:
+        client = boto3.client("s3")
+        f_ranges = {}
+        # Get global min/max map.
+        for i in objects:
+            obj = client.get_object(Bucket=s3_bucket_input, Key=str(i) + "_bounds")["Body"].read()
+            d = json.loads(obj.decode("utf-8"))
+            for idx in d["min"]:
+                v = d["min"][idx]
+                if idx not in f_ranges:
+                    f_ranges[idx] = [v, v]
+                if v < f_ranges[idx][0]:
+                    f_ranges[idx][0] = v
+            for idx in d["max"]:
+                v = d["max"][idx]
+                if idx not in f_ranges:
+                    f_ranges[idx] = [v, v]
+                if v > f_ranges[idx][1]:
+                    f_ranges[idx][1] = v
 
-    client = boto3.client("s3")
-    f_ranges = {}
-    # Get global min/max map.
-    for i in objects:
-        obj = client.get_object(Bucket=s3_bucket_input, Key=str(i) + "_bounds")["Body"].read()
-        d = json.loads(obj.decode("utf-8"))
-        for idx in d["min"]:
-            v = d["min"][idx]
-            if idx not in f_ranges:
-                f_ranges[idx] = [v, v]
-            if v < f_ranges[idx][0]:
-                f_ranges[idx][0] = v
-        for idx in d["max"]:
-            v = d["max"][idx]
-            if idx not in f_ranges:
-                f_ranges[idx] = [v, v]
-            if v > f_ranges[idx][1]:
-                f_ranges[idx][1] = v
-
-    end_global = time.time()
-    print("Creating the global map took {0} seconds...".format(end_global - start_global))
-    # Update local min/max maps.
-    for i in objects:
-        s3_obj = s3_resource.Object(s3_bucket_input, str(i) + "_bounds")
-        obj = s3_obj.get()["Body"].read()
-        d = json.loads(obj.decode("utf-8"))
-        for idx in d["min"]:
-            d["min"][idx] = f_ranges[idx][0]
-        for idx in d["max"]:
-            d["max"][idx] = f_ranges[idx][1]
-        s = json.dumps(d)
-        client.put_object(Bucket=s3_bucket_input, Key=str(i) + "_final_bounds", Body=s)
-        s3_obj.delete()
+        end_global = time.time()
+        print("Creating the global map took {0} seconds...".format(end_global - start_global))
+        # Update local min/max maps.
+        for i in objects:
+            s3_obj = s3_resource.Object(s3_bucket_input, str(i) + "_bounds")
+            obj = s3_obj.get()["Body"].read()
+            d = json.loads(obj.decode("utf-8"))
+            for idx in d["min"]:
+                d["min"][idx] = f_ranges[idx][0]
+            for idx in d["max"]:
+                d["max"][idx] = f_ranges[idx][1]
+            s = json.dumps(d)
+            client.put_object(Bucket=s3_bucket_input, Key=str(i) + "_final_bounds", Body=s)
+            s3_obj.delete()
 
     start_scale = time.time()
     print("Putting local maps took {0} seconds...".format(start_scale - end_global))
@@ -95,7 +103,7 @@ def MinMaxScaler(s3_bucket_input, s3_bucket_output, lower, upper, objects=[], dr
             if len(g_threads) > 400:
                 t = g_threads.popleft()
                 t.join()
-            g = LocalScale(s3_bucket_input, i, s3_bucket_output, lower, upper)
+            g = LocalScale(s3_bucket_input, i, s3_bucket_output, lower, upper, redis)
             g.start()
             g_threads.append(g)
 
@@ -103,6 +111,9 @@ def MinMaxScaler(s3_bucket_input, s3_bucket_output, lower, upper, objects=[], dr
             t.join()
     end_scale = time.time()
     print("Local scaling took {0} seconds...".format(end_scale - start_scale))
+
+    if redis:
+        return
 
     for i in objects:
         s3_resource.Object(s3_bucket_input, str(i) + "_final_bounds").delete()
